+++ packages/assets/fatin-verse-assets/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @avatune/fatin-verse-assets
 
+## 3.1.0
+
+### Minor Changes
+
+- [`064041d`](https://github.com/avatune/avatune/commit/064041dc055e383e53be29b69be5ea5fe2dc6598) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix workspace package version is not resolved
+
+## 3.0.2
+
+### Patch Changes
+
+- [`f5ced92`](https://github.com/avatune/avatune/commit/f5ced920a22ab77bae7fbf50f11194482c0354d6) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix solidjs SSR support
+
 ## 3.0.1
 
 ### Patch Changes