+++ packages/predictors/hair-color-predictor/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @avatune/hair-color-predictor
 
+## 1.4.0
+
+### Minor Changes
+
+- [#81](https://github.com/avatune/avatune/pull/81) [`37dd838`](https://github.com/avatune/avatune/commit/37dd838f32a92ca8ad2af5ae3a62be47040820c3) Thanks [@teimurjan](https://github.com/teimurjan)! - Improve prediction quality with MobileNetV2
+
+## 1.3.0
+
+### Minor Changes
+
+- [`064041d`](https://github.com/avatune/avatune/commit/064041dc055e383e53be29b69be5ea5fe2dc6598) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix workspace package version is not resolved
+
+### Patch Changes
+
+- Updated dependencies [[`064041d`](https://github.com/avatune/avatune/commit/064041dc055e383e53be29b69be5ea5fe2dc6598)]:
+  - @avatune/types@2.2.0
+  - @avatune/face-detector@1.3.0
+
 ## 1.2.6
 
 ### Patch Changes