+++ packages/themes/miniavs-theme/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @avatune/miniavs-theme
 
+## 2.1.0
+
+### Minor Changes
+
+- [`064041d`](https://github.com/avatune/avatune/commit/064041dc055e383e53be29b69be5ea5fe2dc6598) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix workspace package version is not resolved
+
+### Patch Changes
+
+- Updated dependencies [[`064041d`](https://github.com/avatune/avatune/commit/064041dc055e383e53be29b69be5ea5fe2dc6598)]:
+  - @avatune/miniavs-assets@2.1.0
+  - @avatune/theme-builder@1.3.0
+  - @avatune/types@2.2.0
+  - @avatune/utils@1.5.0
+
+## 2.0.4
+
+### Patch Changes
+
+- [`f5ced92`](https://github.com/avatune/avatune/commit/f5ced920a22ab77bae7fbf50f11194482c0354d6) Thanks [@teimurjan](https://github.com/teimurjan)! - Fix solidjs SSR support
+
+- Updated dependencies [[`f5ced92`](https://github.com/avatune/avatune/commit/f5ced920a22ab77bae7fbf50f11194482c0354d6)]:
+  - @avatune/miniavs-assets@2.0.2
+
 ## 2.0.3
 
 ### Patch Changes